teacher.java: change int Id to Long Id

+++ src/main/java/com/project/app/model/Student.java
@@ -3,7 +3,7 @@ package com.project.app.model;
 public class Student extends Human {
 
 
-    public Student(int id, int age, String sex, double salary, String type) {
+    public Student(Long id, int age, String sex, double salary, String type) {
         super(id, age, sex, salary, type);
     }
 }
