+++ src/main/java/com/project/app/model/Teacher.java
@@ -4,7 +4,7 @@ package com.project.app.model;
 
 public class Teacher extends Human {
 
-    public Teacher(int id, int age, String sex, double salary , String type) {
+    public Teacher(Long id, int age, String sex, double salary , String type) {
         super(id,age,sex,salary, type);
     }
 }
